backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
flask-sqlalchemy==3.0.3
apscheduler==3.10.1
python-dotenv==1.0.0
requests==2.28.2
jinja2==3.1.2
urllib3==1.26.15
gunicorn==20.1.0